Split functions

+++ src/Pool/Modules/PoolNonTv.sol
@@ -30,7 +30,7 @@ contract PoolNonTv is Pool {
     constructor(
         address base_,
         address fyToken_,
-        uint128 ts_,
+        uint256 ts_,
         uint16 g1Fee_
     ) Pool(base_, fyToken_, ts_, g1Fee_) {}
